pokerlab
Module hand

Module for poker hand evaluation and comparison.

There is a bit of Poker lingo that this file uses. If you find yourself confused by anything, please reach out to us and ask! The term that will probably be most confusing is a “kicker”. A “kicker”, also known as a “side card”, is simply a card that does not take part in determine the rank of a hand. For example, if we want to compare 2 four-of-a-kinds, the fifth card in each hand would be considered the kicker. Note that kickers are still important in breaking ties.

A good way to approach this assignment is to first read the code in src/card.rs, and then read all of the starter code in this file (src/hand.rs). If you are unfamiliar with Poker, you may also want to read up on the hands you can make on Wikipedia. Finally (as a good rule of thumb for any programming course), make sure to go over the test cases to make sure you have a correct idea of what you should implement.

You are also allowed to change any of the private fields of structs in this file, as long as you are able to implement the all of the Hand methods without issue. You are also allowed to modify the PokerHand::solve method, but be aware that the test cases are solely using that method to check the correctness of your implementation, so make sure not to play around with it too much.

Structs

Flush

Represents five cards of the same suit. The suit of the cards is ignored since we do not need it for comparison.

FourOfAKind

Represents a hand containing four cards of the same rank and one kicker.

FullHouse

Represents a hand containing three cards of one rank and two cards of another rank.

Hand

Represents a standard Hand of 5 playing Cards.

HighCard

Represents a high card hand, consisting of five unpaired cards.

OnePair

Represents a hand containing one pair and three kickers.

Straight

Represents five consecutive cards of different suits, storing only the highest card rank.

StraightFlush

Represents five consecutive cards of the same suit. The suit of the cards is ignored since we do not need it for comparison.

ThreeOfAKind

Represents a hand containing three cards of the same rank and two kickers.

TwoPair

Represents a hand containing two pairs and one kicker.

Enums

PokerHand

Represents different poker hand rankings with their respective cards. Each variant contains the relevant cards that make up the hand.