pokerlab

Module card

This module contains the definition and implementation of Card, which represents a card in a standard deck of 52 cards. Cards are uniquely identified by their rank (which can be the numbers 2-10 or Jack, Queen, King, or Ace) and their suit (which can be Diamond, Club, Heart, or Spade).

Note to students: You are allowed to modify this file and replace the implementation with something similar to your own implementation in Card Lab.

However, make sure you understand that the function signatures are different, and that the Rank and Suit types are now public! Additionally, we now no longer make a distinction between suits. For example, we will say that the Ace of Spades has equal value to the Ace of Diamonds.

Structs

Card

Represents a standard playing card with a suit and a rank.

Enums

Rank

Represents the rank of a playing card, which can be either a number card (2-10) or a face card (Ace, King, Queen, Jack).

Suit

Represents the four possible suits in a standard deck of playing cards.